requirements.txt
sentence-transformers
faiss-cpu
numpy
langchain
langchain-openai
openai
python-dotenv
gitpython
streamlit

# evaluation
ragas
datasets

# experiment tracking
mlflow

rank-bm25